requirements.txt
# SMS Service Dependencies

# Twilio SDK for sending SMS
twilio>=8.10.0

# Environment variable management
python-dotenv>=1.0.0

# Retry logic with exponential backoff
tenacity>=8.2.0

# Testing (optional, for development)
pytest>=7.4.0
pytest-mock>=3.12.0
